requirements.txt
openai
torch
PyPDF2
ollama
pdf2image
pytesseract
pdf2image
Pillow
tkinter
pytesseract
spellchecker
pyspellchecker

